finished moving some more things
